✨configurable offset for all output 'date' field
Source measures are timestamped with the time
of the measure which covers the interval from
the last one. But the interval of the measure
starts with the previous measure time.
Because of that old P5D, F5D... output 'data',
was set to the start of the interval (1h less).
This criterium has been applied to all the curves.
Since P2 use quarter hour intervals, interval is
a parameter that changes for those curves.

Co-authored-by: javikalsan <jcalzadosanchez@gmail.com>